Refactor asset downloading and app running in launcher
The changes include refactoring the way assets are downloaded in the launcher program. Now, instead of downloading individual .exe and .json files, a single asset (zip file) is downloaded, unzipped and then used in the app. Directory cleanup functionality is also added before downloading the zip file. The path for the app's executable file is adjusted to reflect the changes.