Update GetCurrentVersion method to use process info
The logic in the GetCurrentVersion method was updated to retrieve the version number directly from the executable file. This is done by creating a process, attaching it to the executable file, and extracting the file version information from the process. This method is more reliable and dynamic than hardcoding the version in the method.