Improve buffer height adjustment process in Console
This commit adds more robust handling for adjusting the Console buffer height in the AppUtilities file. If the desired buffer height is too large, the code will now attempt to set it to the maximum allowed value. If this is too large, it will then set it to a default value, preventing potential ArgumentOutOfRangeExceptions.