Removed infinite loop in MouseMove and Notepad dependency
In the RandomMouseMover class, the movable area of the mouse cursor within the screen has been reduced from a constant loop to a single movement. This removed the constant thread sleep, and now the movement will happen whenever invoked from the main Program class.
In addition, removed all dependencies on Notepad from the KeyboardSimulator and AppUtilities classes. Now the key press will occur in the console or whichever application is currently active and does not check if Notepad is in focus.